Another big refactor
The main objective here was to make the presence piece more OO and to separate it from the tray icon logic so that this can all run without an icon. This (theoretically) allows non-Windows sytems to run the app.

Also added a bit of retry logic to the RPC calls based on an error seen after the app was left running for a long time.